Improve tests for S2637 and move to compiled module
