+++ java-checks-test-sources/src/main/java/symbolicexecution/checks/NonNullSetToNullCheck/packageNonNull/package-info.java
@@ -1,0 +1,2 @@
+@org.eclipse.jdt.annotation.NonNullByDefault
+package symbolicexecution.checks.NonNullSetToNullCheck.packageNonNull;
